agent-platform/pyproject.toml
[project]
name = "agentforge"
version = "0.1.0"
description = "A lightweight, self-hostable agentic AI platform"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=12.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "openai>=1.12.0",
    "anthropic>=0.21.0",
    "chromadb>=0.4.24",
    "beautifulsoup4>=4.12.0",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "watchdog>=4.0.0",
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3.0",
]
faiss = [
    "faiss-cpu>=1.8.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["core*", "llm*", "tools*", "agents*", "api*"]
exclude = ["ui*", "configs*", "examples*"]

[tool.ruff]
line-length = 100
target-version = "py310"
